+++ android/src/main/java/com/unact/yandexmapkit/YandexSearchHandlerImpl.java
@@ -4,14 +4,9 @@ import android.content.Context;
 
 import androidx.annotation.NonNull;
 
-import com.yandex.mapkit.GeoObject;
-import com.yandex.mapkit.GeoObjectCollection;
 import com.yandex.mapkit.geometry.Geometry;
 import com.yandex.mapkit.geometry.Point;
 import com.yandex.mapkit.geometry.BoundingBox;
-import com.yandex.mapkit.search.Address;
-import com.yandex.mapkit.search.BusinessObjectMetadata;
-import com.yandex.mapkit.search.Response;
 import com.yandex.mapkit.search.SearchOptions;
 import com.yandex.mapkit.search.Session;
 import com.yandex.mapkit.search.Snippet;
@@ -21,15 +16,14 @@ import com.yandex.mapkit.search.SearchManagerType;
 import com.yandex.mapkit.search.SuggestOptions;
 import com.yandex.mapkit.search.SearchManager;
 import com.yandex.mapkit.search.SuggestSession;
-import com.yandex.mapkit.search.ToponymObjectMetadata;
 import com.yandex.runtime.Error;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
+import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
@@ -38,16 +32,19 @@ import io.flutter.plugin.common.MethodChannel.Result;
 public class YandexSearchHandlerImpl implements MethodCallHandler {
 
   private MethodChannel methodChannel;
-
-  private Map<Integer, SuggestSession> suggestSessionsById = new HashMap<>();
-  private Session searchSession;
-
   private final SearchManager searchManager;
+	private final BinaryMessenger binaryMessenger;
 
-  public YandexSearchHandlerImpl(Context context, MethodChannel channel) {
+  private Map<Integer, SuggestSession>  		suggestSessionsById = new HashMap<>();
+  private Map<Integer, YandexSearchSession> searchSessionsById  = new HashMap<>();;
+
+  public YandexSearchHandlerImpl(Context context, MethodChannel channel, BinaryMessenger messenger) {
+
     SearchFactory.initialize(context);
-    methodChannel = channel;
-    searchManager = SearchFactory.getInstance().createSearchManager(SearchManagerType.COMBINED);
+
+    methodChannel 	= channel;
+    searchManager 	= SearchFactory.getInstance().createSearchManager(SearchManagerType.COMBINED);
+		binaryMessenger	= messenger;
   }
 
   @Override
@@ -62,13 +59,11 @@ public class YandexSearchHandlerImpl implements MethodCallHandler {
         result.success(null);
         break;
       case "searchByText":
-        searchByText(call);
-        result.success(null);
+        searchByText(call, result);
         break;
-      case "cancelSearch":
-        cancelSearch();
-        result.success(null);
-        break;
+			case "searchByPoint":
+				searchByPoint(call, result);
+				break;
       default:
         result.notImplemented();
         break;
@@ -102,107 +97,153 @@ public class YandexSearchHandlerImpl implements MethodCallHandler {
     suggestSessionsById.put(listenerId, suggestSession);
   }
 
-  public void searchByText(MethodCall call) {
+  public void searchByText(MethodCall call, MethodChannel.Result result) {
 
-    Map<String, Object> params = ((Map<String, Object>) call.arguments);
+		Map<String, Object> params = ((Map<String, Object>) call.arguments);
 
-    String              searchText  = (String) params.get("searchText");
-    Map<String, Object> geometry    = (Map<String, Object>) params.get("geometry");
-    Map<String, Object> options     = (Map<String, Object>) params.get("options");
+		int                 sessionId   = ((Number) params.get("sessionId")).intValue();
+		String              searchText  = (String) params.get("searchText");
+		Map<String, Object> geometry    = (Map<String, Object>) params.get("geometry");
+		Map<String, Object> options     = (Map<String, Object>) params.get("options");
 
-    Geometry geometryObj;
+		Geometry geometryObj;
 
-    if (geometry.containsKey("point")) {
+		if (geometry.containsKey("point")) {
 
-      Map<String, Object> point = (Map<String, Object>) geometry.get("point");
+		  Map<String, Object> point = (Map<String, Object>) geometry.get("point");
 
-      geometryObj = Geometry.fromPoint(
-              new Point(((Double) point.get("latitude")), ((Double) point.get("longitude")))
-      );
+		  geometryObj = Geometry.fromPoint(
+		    new Point(((Double) point.get("latitude")), ((Double) point.get("longitude")))
+		  );
 
-    } else {
+		} else {
 
-      Map<String, Object> boundingBox = (Map<String, Object>) geometry.get("boundingBox");
+		  Map<String, Object> boundingBox = (Map<String, Object>) geometry.get("boundingBox");
 
-      Map<String, Object> southWest = (Map<String, Object>) boundingBox.get("southWest");
-      Map<String, Object> northEast = (Map<String, Object>) boundingBox.get("northEast");
+		  Map<String, Object> southWest = (Map<String, Object>) boundingBox.get("southWest");
+		  Map<String, Object> northEast = (Map<String, Object>) boundingBox.get("northEast");
 
-      geometryObj = Geometry.fromBoundingBox(
-        new BoundingBox(
-          new Point(((Double) southWest.get("latitude")), ((Double) southWest.get("longitude"))),
-          new Point(((Double) northEast.get("latitude")), ((Double) northEast.get("longitude")))
-        )
-      );
-    }
+		  geometryObj = Geometry.fromBoundingBox(
+		    new BoundingBox(
+		      new Point(((Double) southWest.get("latitude")), ((Double) southWest.get("longitude"))),
+		      new Point(((Double) northEast.get("latitude")), ((Double) northEast.get("longitude")))
+		    )
+		  );
+		}
 
-    int                 searchTypeOption     = ((Number) options.get("searchType")).intValue();
-    Number              resultPageSizeOption = (Number) options.get("resultPageSize");
-    Map<String, Object> userPositionOption   = (Map<String, Object>) options.get("userPosition");
+		SearchOptions searchOptions = getSearchOptions(options);
 
-    Integer resultPageSize = null;
-    if (resultPageSizeOption != null) {
-      resultPageSize = resultPageSizeOption.intValue();
-    }
+		Session searchSession = searchManager.submit(
+		  searchText,
+		  geometryObj,
+		  searchOptions,
+		  new YandexSearchListener(result, 0)
+		);
 
-    // Theses params are not implemented on the flutter side yet
-    int snippetOption = Snippet.NONE.value;
-    List<String> experimentalSnippetsOption = new ArrayList<>();
+		YandexSearchSession session = new YandexSearchSession(
+			sessionId,
+			searchSession,
+			binaryMessenger,
+			new CloseSearchSessionCallback()
+		);
 
-    Point userPosition = null;
-
-    if (userPositionOption != null) {
-      userPosition = new Point(((Double) userPositionOption.get("latitude")), ((Double) userPositionOption.get("longitude")));
-    }
-
-    String  originOption                    = (String) options.get("origin");
-    String  directPageIdOption              = (String) options.get("directPageId");
-    String  appleCtxOption                  = (String) options.get("appleCtx");
-    Boolean geometryOption                  = (Boolean) options.get("geometry");
-    String  advertPageIdOption              = (String) options.get("advertPageId");
-    Boolean suggestWordsOption              = (Boolean) options.get("suggestWords");
-    Boolean disableSpellingCorrectionOption = (Boolean) options.get("disableSpellingCorrection");
-
-    if (geometryOption == null) {
-      geometryOption = false;
-    }
-
-    if (suggestWordsOption == null) {
-      suggestWordsOption = false;
-    }
-
-    if (disableSpellingCorrectionOption == null) {
-      disableSpellingCorrectionOption = false;
-    }
-
-    SearchOptions searchOptions = new SearchOptions(
-      searchTypeOption,
-      resultPageSize,
-      snippetOption,
-      experimentalSnippetsOption,
-      userPosition,
-      originOption,
-      directPageIdOption,
-      appleCtxOption,
-      geometryOption,
-      advertPageIdOption,
-      suggestWordsOption,
-      disableSpellingCorrectionOption
-    );
-
-    searchSession = searchManager.submit(
-      searchText,
-      geometryObj,
-      searchOptions,
-      new YandexSearchListener()
-    );
+		searchSessionsById.put(sessionId, session);
   }
 
-  public void cancelSearch() {
+	public void searchByPoint(MethodCall call, MethodChannel.Result result) {
 
-    if (searchSession != null) {
-      searchSession.cancel();
-    }
+		Map<String, Object> params = ((Map<String, Object>) call.arguments);
+
+		int 								sessionId	= ((Number) params.get("sessionId")).intValue();
+		Map<String, Object> point    	= (Map<String, Object>) params.get("point");
+		Integer             zoom  		= (Integer) params.get("zoom");
+		Map<String, Object> options   = (Map<String, Object>) params.get("options");
+
+		SearchOptions searchOptions = getSearchOptions(options);
+
+		Session searchSession = searchManager.submit(
+			new Point(((Double) point.get("latitude")), ((Double) point.get("longitude"))),
+			zoom,
+			searchOptions,
+			new YandexSearchListener(result, 0)
+		);
+
+		YandexSearchSession session = new YandexSearchSession(
+			sessionId,
+			searchSession,
+			binaryMessenger,
+			new CloseSearchSessionCallback()
+		);
+
+		searchSessionsById.put(sessionId, session);
+	}
+
+  private SearchOptions getSearchOptions(Map<String, Object> options) {
+
+		int                 searchTypeOption     = ((Number) options.get("searchType")).intValue();
+		Number              resultPageSizeOption = (Number) options.get("resultPageSize");
+		Map<String, Object> userPositionOption   = (Map<String, Object>) options.get("userPosition");
+
+		Integer resultPageSize = null;
+		if (resultPageSizeOption != null) {
+		  resultPageSize = resultPageSizeOption.intValue();
+		}
+
+		// Theses params are not implemented on the flutter side yet
+		int snippetOption = Snippet.NONE.value;
+		List<String> experimentalSnippetsOption = new ArrayList<>();
+
+		Point userPosition = null;
+
+		if (userPositionOption != null) {
+		  userPosition = new Point(((Double) userPositionOption.get("latitude")), ((Double) userPositionOption.get("longitude")));
+		}
+
+		String  originOption                    = (String) options.get("origin");
+		String  directPageIdOption              = (String) options.get("directPageId");
+		String  appleCtxOption                  = (String) options.get("appleCtx");
+		Boolean geometryOption                  = (Boolean) options.get("geometry");
+		String  advertPageIdOption              = (String) options.get("advertPageId");
+		Boolean suggestWordsOption              = (Boolean) options.get("suggestWords");
+		Boolean disableSpellingCorrectionOption = (Boolean) options.get("disableSpellingCorrection");
+
+		if (geometryOption == null) {
+		  geometryOption = false;
+		}
+
+		if (suggestWordsOption == null) {
+		  suggestWordsOption = false;
+		}
+
+		if (disableSpellingCorrectionOption == null) {
+		  disableSpellingCorrectionOption = false;
+		}
+
+		SearchOptions searchOptions = new SearchOptions(
+		  searchTypeOption,
+		  resultPageSize,
+		  snippetOption,
+		  experimentalSnippetsOption,
+		  userPosition,
+		  originOption,
+		  directPageIdOption,
+		  appleCtxOption,
+		  geometryOption,
+		  advertPageIdOption,
+		  suggestWordsOption,
+		  disableSpellingCorrectionOption
+		);
+
+		return searchOptions;
   }
+
+	private class CloseSearchSessionCallback implements YandexSearchSessionCloseCallbackInterface {
+
+		@Override
+		public void onClose(int sessionId) {
+			searchSessionsById.remove(sessionId);
+		}
+	}
 
   private class YandexSuggestListener implements SuggestSession.SuggestListener {
 
@@ -241,260 +282,9 @@ public class YandexSearchHandlerImpl implements MethodCallHandler {
 
     @Override
     public void onError(@NonNull Error error) {
-      Map<String, Object> arguments = new HashMap<>();
-      arguments.put("listenerId", listenerId);
-      methodChannel.invokeMethod("onSuggestListenerError", arguments);
-    }
-  }
-
-  private class YandexSearchListener implements Session.SearchListener {
-
-    @Override
-    public void onSearchResponse(@NonNull Response response) {
-
-      Map<String, Object> data = new HashMap<>();
-
-      data.put("found", response.getMetadata().getFound());
-
-      List<Map<String, Object>> dataItems = new ArrayList<>();
-
-      Iterator<GeoObjectCollection.Item> objectsIterator = response.getCollection().getChildren().iterator();
-
-      while (objectsIterator.hasNext()) {
-
-        GeoObjectCollection.Item item = objectsIterator.next();
-
-        GeoObject obj = item.getObj();
-        if (item.getObj() == null) {
-          continue;
-        }
-
-        Map<String, Object> dataItem = new HashMap<>();
-
-        dataItem.put("name", obj.getName());
-
-        List<Map<String, Object>> geometry = new ArrayList<>();
-
-        Iterator<Geometry> geometryIterator = obj.getGeometry().iterator();
-
-        while (geometryIterator.hasNext()) {
-
-          final Geometry geometryItem = geometryIterator.next();
-
-          if (geometryItem.getPoint() != null) {
-            geometry.add(
-              new HashMap<String,Object>() {{
-                put(
-                  "point",
-                  new HashMap<String, Object>() {{
-                    put("latitude", geometryItem.getPoint().getLatitude());
-                    put("longitude", geometryItem.getPoint().getLongitude());
-                  }}
-                );
-              }}
-            );
-          }
-
-          if (geometryItem.getBoundingBox() != null) {
-            geometry.add(
-              new HashMap<String,Object>() {{
-                put("boundingBox",
-                  new HashMap<String, Object>() {{
-                    put("southWest", new HashMap<String, Object>() {{
-                      put("latitude", geometryItem.getBoundingBox().getSouthWest().getLatitude());
-                      put("longitude", geometryItem.getBoundingBox().getSouthWest().getLongitude());
-                    }});
-                    put("northEast", new HashMap<String, Object>() {{
-                      put("latitude", geometryItem.getBoundingBox().getNorthEast().getLatitude());
-                      put("longitude", geometryItem.getBoundingBox().getNorthEast().getLongitude());
-                    }});
-                  }}
-                );
-              }}
-            );
-          }
-        }
-
-        dataItem.put("geometry", geometry);
-
-        ToponymObjectMetadata toponymMeta = obj.getMetadataContainer().getItem(ToponymObjectMetadata.class);
-        if (toponymMeta != null) {
-          dataItem.put("toponymMetadata", getToponymMetadata(toponymMeta));
-        }
-
-        BusinessObjectMetadata businessMeta = obj.getMetadataContainer().getItem(BusinessObjectMetadata.class);
-        if (businessMeta != null) {
-          dataItem.put("businessMetadata", getBusinessMetadata(businessMeta));
-        }
-
-        dataItems.add(dataItem);
-      }
-
-      data.put("items", dataItems);
-
-      Map<String, Object> arguments = new HashMap<>();
-
-      arguments.put("response", data);
-
-      methodChannel.invokeMethod("onSearchListenerResponse", arguments);
-    }
-
-    private Map<String, Object> getToponymMetadata(ToponymObjectMetadata meta) {
-
-      Map<String, Object> toponymMetadata = new HashMap<>();
-
-      Map<String, Double> balloonPoint = new HashMap<>();
-
-      balloonPoint.put("latitude", meta.getBalloonPoint().getLatitude());
-      balloonPoint.put("longitude", meta.getBalloonPoint().getLongitude());
-
-      toponymMetadata.put("balloonPoint", balloonPoint);
-
-      Map<String, Object> address = new HashMap<>();
-
-      address.put("formattedAddress", meta.getAddress().getFormattedAddress());
-      address.put("addressComponents", getAddressComponents(meta.getAddress()));
-
-      toponymMetadata.put("address", address);
-
-      return toponymMetadata;
-    }
-
-    private Map<String, Object> getBusinessMetadata(BusinessObjectMetadata meta) {
-
-      Map<String, Object> businessMetadata = new HashMap<>();
-
-      businessMetadata.put("name", meta.getName());
-
-      if (meta.getShortName() != null) {
-        businessMetadata.put("shortName", meta.getShortName());
-      }
-
-      Map<String, Object> address = new HashMap<>();
-
-      address.put("formattedAddress", meta.getAddress().getFormattedAddress());
-      address.put("addressComponents", getAddressComponents(meta.getAddress()));
-
-      businessMetadata.put("address", address);
-
-      return businessMetadata;
-    }
-
-    private Map<Integer, String> getAddressComponents(Address address) {
-
-      Map<Integer, String> addressComponents = new HashMap<>();
-
-      Iterator<Address.Component> iterator = address.getComponents().iterator();
-
-      while (iterator.hasNext()) {
-
-        Address.Component addressComponent = iterator.next();
-
-        Integer flutterKind = 0;
-
-        String value = addressComponent.getName();
-
-        Iterator<Address.Component.Kind> addressKindsIterator = addressComponent.getKinds().iterator();
-
-        while (addressKindsIterator.hasNext()) {
-
-          Address.Component.Kind addressComponentKind = addressKindsIterator.next();
-
-          switch (addressComponentKind) {
-            case UNKNOWN:
-              flutterKind = 0;
-              break;
-            case COUNTRY:
-              flutterKind = 2;
-              break;
-            case REGION:
-              flutterKind = 3;
-              break;
-            case PROVINCE:
-              flutterKind = 4;
-              break;
-            case AREA:
-              flutterKind = 5;
-              break;
-            case LOCALITY:
-              flutterKind = 6;
-              break;
-            case DISTRICT:
-              flutterKind = 7;
-              break;
-            case STREET:
-              flutterKind = 8;
-              break;
-            case HOUSE:
-              flutterKind = 9;
-              break;
-            case ENTRANCE:
-              flutterKind = 10;
-              break;
-            case ROUTE:
-              flutterKind = 11;
-              break;
-            case STATION:
-              flutterKind = 12;
-              break;
-            case METRO_STATION:
-              flutterKind = 13;
-              break;
-            case RAILWAY_STATION:
-              flutterKind = 14;
-              break;
-            case VEGETATION:
-              flutterKind = 15;
-              break;
-            case HYDRO:
-              flutterKind = 16;
-              break;
-            case AIRPORT:
-              flutterKind = 17;
-              break;
-            case OTHER:
-              break;
-          }
-
-          addressComponents.put(flutterKind, value);
-        }
-      }
-
-      return addressComponents;
-    }
-
-    @Override
-    public void onSearchError(@NonNull Error error) {
-
-      Map<String, Object> arguments = new HashMap<>();
-
-      String errorMessage = "Unknown error";
-
-      if (error instanceof SearchError) {
-        errorMessage = ((SearchError) error).getMessage();
-      }
-
-      arguments.put("error", errorMessage);
-
-      methodChannel.invokeMethod("onSearchListenerError", arguments);
-    }
-  }
-
-  private class SearchError implements Error {
-
-    private String message;
-
-    public SearchError(String message) {
-      this.message = message;
-    }
-
-    public String getMessage() {
-      return message;
-    }
-
-    @Override
-    public boolean isValid() {
-      return false;
+			Map<String, Object> arguments = new HashMap<>();
+			arguments.put("listenerId", listenerId);
+			methodChannel.invokeMethod("onSuggestListenerError", arguments);
     }
   }
 }
